LayoutTests/fast/canvas/imagedata-storageformat-enabled-expected.txt

Tests that ImageDataSettings contains storageFormat.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS created_imageData_uint8.data.constructor is Uint8ClampedArray
PASS gotten_imageData_uint8.data.constructor is Uint8ClampedArray
PASS created_imageData_float16.data.constructor is Uint8ClampedArray
PASS gotten_imageData_float16.data.constructor is Uint8ClampedArray
PASS context.createImageData(1, 1, { storageFormat: "foo" }) threw exception TypeError: Type error.
PASS context.getImageData(0, 0, 1, 1, { storageFormat: "foo" }) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
